LPS-64277 Rename package
