+++ modules/apps/knowledge-base/markdown-converter/src/main/java/com/liferay/knowledge/base/markdown/converter/internal/pegdown/ast/PicWithCaptionNode.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.markdown.converter.internal.pegdown.ast;
+package com.liferay.knowledge.base.markdown.converter.internal.pegdown.ast;
 
 import java.util.List;
